Simplify range action implementations + corrections after review
