+++ data/crac/crac-impl/src/main/java/com/farao_community/farao/data/crac_impl/json/serializers/range_action/RangeActionSerializer.java
@@ -7,6 +7,7 @@
 
 package com.farao_community.farao.data.crac_impl.json.serializers.range_action;
 
+import com.farao_community.farao.data.crac_api.ExtensionsHandler;
 import com.farao_community.farao.data.crac_api.RangeAction;
 import com.farao_community.farao.data.crac_impl.json.JsonSerializationNames;
 import com.farao_community.farao.data.crac_impl.json.serializers.AbstractRemedialActionSerializer;
@@ -17,13 +18,16 @@ import com.fasterxml.jackson.core.JsonToken;
 import com.fasterxml.jackson.core.type.WritableTypeId;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.jsontype.TypeSerializer;
+import com.powsybl.commons.json.JsonUtil;
 
 import java.io.IOException;
+import java.util.Optional;
 
 /**
  * @author Joris Mancini {@literal <joris.mancini at rte-france.com>}
  */
 public class RangeActionSerializer<E extends AbstractRangeAction> extends AbstractRemedialActionSerializer<RangeAction, E> {
+
     @Override
     public void serialize(E abstractRangeAction, JsonGenerator jsonGenerator, SerializerProvider serializerProvider) throws IOException {
         super.serialize(abstractRangeAction, jsonGenerator, serializerProvider);
@@ -32,6 +36,11 @@ public class RangeActionSerializer<E extends AbstractRangeAction> extends Abstra
             jsonGenerator.writeObject(range);
         }
         jsonGenerator.writeEndArray();
+        Optional<String> groupId = abstractRangeAction.getGroupId();
+        if (groupId.isPresent()) {
+            jsonGenerator.writeStringField("groupId", groupId.get());
+        }
+        JsonUtil.writeExtensions(abstractRangeAction, jsonGenerator, serializerProvider, ExtensionsHandler.getExtensionsSerializers());
     }
 
     @Override
