--- data/crac/crac-impl/src/main/java/com/farao_community/farao/data/crac_impl/remedial_action/range_action/Countertrading.java
@@ -1,83 +1,0 @@
-/*
- * Copyright (c) 2019, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-
-package com.farao_community.farao.data.crac_impl.remedial_action.range_action;
-
-import com.farao_community.farao.data.crac_api.NetworkElement;
-import com.farao_community.farao.data.crac_api.usage_rule.UsageRule;
-import com.farao_community.farao.data.crac_impl.range_domain.Range;
-import com.fasterxml.jackson.annotation.JsonTypeName;
-import com.powsybl.iidm.network.Network;
-
-import java.util.HashSet;
-import java.util.List;
-import java.util.Optional;
-import java.util.Set;
-
-/**
- * Countertrading remedial action.
- *
- * @author Viktor Terrier {@literal <viktor.terrier at rte-france.com>}
- */
-@JsonTypeName("countertrading")
-public class Countertrading extends AbstractRangeAction {
-
-    public static final int TEMP_VALUE = 0;
-
-    public Countertrading(String id, String name, String operator, List<UsageRule> usageRules, List<Range> ranges) {
-        super(id, name, operator, usageRules, ranges);
-    }
-
-    public Countertrading(String id, String name, String operator) {
-        super(id, name, operator);
-    }
-
-    @Override
-    public double getMinValue(Network network) {
-        return TEMP_VALUE;
-    }
-
-    @Override
-    public double getMaxValue(Network network) {
-        return TEMP_VALUE;
-    }
-
-    @Override
-    public double getCurrentValue(Network network) {
-        return TEMP_VALUE;
-    }
-
-    @Override
-    public void apply(Network network, double setpoint) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Optional<String> getGroupId() {
-        return Optional.empty();
-    }
-
-    @Override
-    public Set<NetworkElement> getNetworkElements() {
-        return new HashSet<>();
-    }
-
-    @Override
-    public void synchronize(Network network) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void desynchronize() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public boolean isSynchronized() {
-        throw new UnsupportedOperationException();
-    }
-}
