+++ data/crac/crac-impl/src/main/java/com/farao_community/farao/data/crac_impl/remedial_action/range_action/HvdcRange.java
@@ -22,7 +22,7 @@ import java.util.List;
  * @author Viktor Terrier {@literal <viktor.terrier at rte-france.com>}
  */
 @JsonTypeName("hvdc-range")
-public final class HvdcRange extends AbstractElementaryRangeAction {
+public final class HvdcRange extends AbstractRangeAction {
 
     protected static int hvdcRangeTempValue = 0;
