+++ data/crac/crac-impl/src/main/java/com/farao_community/farao/data/crac_impl/remedial_action/range_action/InjectionRange.java
@@ -21,7 +21,7 @@ import java.util.List;
  * @author Viktor Terrier {@literal <viktor.terrier at rte-france.com>}
  */
 @JsonTypeName("injection-range")
-public final class InjectionRange extends AbstractElementaryRangeAction {
+public final class InjectionRange extends AbstractRangeAction {
 
     protected static int injectionRangeTempValue = 0;
