+++ data/crac/crac-impl/src/main/java/com/farao_community/farao/data/crac_impl/remedial_action/range_action/Redispatching.java
@@ -23,7 +23,7 @@ import java.util.Set;
  * @author Viktor Terrier {@literal <viktor.terrier at rte-france.com>}
  */
 @JsonTypeName("redispatching")
-public final class Redispatching extends AbstractElementaryRangeAction {
+public final class Redispatching extends AbstractRangeAction {
 
     public static final int TEMP_VALUE_REDISPATCH = 0;
