+++ data/crac/crac-impl/src/test/java/com/farao_community/farao/data/crac_impl/remedial_action/range_action/HvdcRangeTest.java
@@ -25,7 +25,7 @@ import static org.junit.Assert.*;
 /**
  * @author Alexandre Montigny {@literal <alexandre.montigny at rte-france.com>}
  */
-public class HvdcRangeTest extends AbstractElementaryRangeActionTest {
+public class HvdcRangeTest extends AbstractRangeActionTest {
 
     private HvdcRange hvdcRange;
