+++ data/crac/crac-impl/src/test/java/com/farao_community/farao/data/crac_impl/remedial_action/range_action/InjectionRangeTest.java
@@ -25,7 +25,7 @@ import static org.junit.Assert.*;
 /**
  * @author Alexandre Montigny {@literal <alexandre.montigny at rte-france.com>}
  */
-public class InjectionRangeTest extends AbstractElementaryRangeActionTest {
+public class InjectionRangeTest extends AbstractRangeActionTest {
 
     private InjectionRange injectionRange;
