+++ data/crac/crac-impl/src/test/java/com/farao_community/farao/data/crac_impl/remedial_action/range_action/PstWithRangeTest.java
@@ -27,7 +27,7 @@ import static org.mockito.Mockito.spy;
 /**
  * @author Joris Mancini {@literal <joris.mancini at rte-france.com>}
  */
-public class PstWithRangeTest extends AbstractElementaryRangeActionTest {
+public class PstWithRangeTest extends AbstractRangeActionTest {
     private String networkElementId;
     private PstWithRange pst;
     private Network network;
